Refactor generate route code files
